Aggiustato un long in users

+++ src/main/java/com/example/EasyBlog/Controller/ArticlesController.java
@@ -40,7 +40,7 @@ public class ArticlesController {
     public Optional<List<Articles>> getArticlesByGender(@PathVariable TypeGenderArticlesEnum typeGenderArticlesEnum){return articlesService.getArticlesByGender(typeGenderArticlesEnum);}
 
     @PutMapping("/{id}/updateStatus")
-    public Optional<Articles> updateArticle(@PathVariable Integer id){return articlesService.updateArticleStatus(id);}
+    public Optional<Articles> updateArticle(@PathVariable Long id){return articlesService.updateArticleStatus(id);}
 
     @PutMapping("/{id}/update")
     public ResponseEntity<Articles> updateArticle(@PathVariable Long id, @RequestBody Articles articles) {
